use IMMEDIATE transaction mode for put_block
put_block performs a large transaction which was started with the
default settings. This resulted in a behavior, when a put_block
transaction, initially holding SHARED, tried to promote itself to
RESERVED or EXCLUSIVE. Such attempt may fail due to another transaction
holding either type of write lock. If the error occurs in the middle of
a transation, busy_handler is bypassed result resulting in SQLITE_BUSY
error, while the error in the beginning of transaction allows to make
use of busy_handler.