Run fmt and lints in the same job with tests
cargo clippy also wants to build dependencies, so it should benefit
from the cache.

As cargo fmt remained the only test step in its own parallel job,
it does not seem sensible to install a toolchain separately just for
this step, as it adds only a few seconds to the job budget.