Add GitHub workflow to update Cargo.lock
Improve caching in the CI by maintaining Cargo.lock checked into
the single commit in ci/cargo-lock branch that is automatically rebased
against master.
At each push to master, the ci/cargo-lock branch is grafted onto the
head commit, then cargo update is run. If Cargo.lock is modified, its
ci/cargo-lock branch commit is reformed with the new content and
the branch is force-pushed to GitHub.

The test workflow then cherry-picks the Cargo.lock onto the
revision under test, and uses a hash of its contents to get the precise
hit in the cargo registry cache in the most common case when the
baseline Cargo.lock is up to date with regard to the build.