Collate Cargo.toml files for CircleCI cache key
The previously used key only accounted for the Cargo.toml at the root
of the workspace tree. To get good cache coherence, we need to checksum
all Cargo.toml files in a stable order.
No need to change the cache key version, because the structure of
the source information remains the same, the content will just
get different hits.