Improve CircleCI cache coherence
Do this by splitting the cargo caches into the crates.io index cache
and the dependency crate cache, which also no longer redundantly stores
the sources. The index cache key varies with the hash of the head commit
in the index repository. The index cache is utilised and updated
by the cargo generate-lockfile command to produce Cargo.lock, which is
used to get a perfect match in the crate cache.