Rework server handshake and client_auth methods
Change the nonce parameter signature to pass a byte slice instead of a
Vec by value.

Pass the peer address to handshake and client_auth methods of the
Node service trait.
Just like with subscription requests, this is needed to
identify the connection for the stateless service.

Make both methods async.