Add node ID crypto into chain-network
Provide the methods and types necessary for node ID auth in
chain-network, using chain-crypto under the hood.

Add NodeKeyPair which encapsulates an Ed25519 key pair and
has methods to generate the key pair and provide an
AuthenticatedNodeId for the provided nonce.

Document node ID/auth data types and methods in data::p2p.