Make private vote talling fallible
When the discrete log search for the private choice's weight fails to
find the result, this should be assumed to be an error in the inputs:
either the max_votes argument is picked wrongly, or any of the
state or decrypt shares give erroneous data. Assuming no programming
error with choice of max_votes, we need to return an error rather than
converting the result to 0.