sparse-array: proptest usage proof-of-concept
This demonstrates the basic use of `proptest` instead of `quickcheck`
including examples of high-order strategies.

`proptest!` and `prop_compose!` macros are not used **on purpose**. They
have their own limitations and the syntax inside them is not valid Rust
syntex. On top of that, they are `cargo fmt`-friendly. With that in
mind, it is preferred to abstain from using these macros.

One thing that could be improved is the way `TestRunner` is used.
Instead of using `proptest!` or invoking the runner directly, an
attribute-style macro could be implemented (this is not set in stone,
just an example of how it may look like):

    #[proptest(strategy = "strategy to generate Value")]
    fn test_case(v: Value) -> TestCaseResult {
        // the actual test code goes here
    }

This is perfectly valid Rust code and it is less boilerplaty than
invoking `TestRunner` by hand.