Ristretto flag (#555)
Resolved a non-compiling feature. Changed the backend of ristretto from zerocaf to curve-25519-dalek-dg.

* Removed zerocaf and added ristretto255
* Improved benches

note: removed the normalization of the results, as there is no clear use for it. The to_bytes() fn already takes care of normalizing the points, and eq() fn does not depend on normalized or non-normalized. 

Co-authored-by: Giacomo Pasini <giacomo.pasini@iohk.io>