forbid voting with multisig accounts
And use Identifier instead of UnspecifiedAccountIdentifier where
necessary. Currently there is nothing stopping voting with multisig
accounts, but it's not possible to mint tokens for those, and we are not
considering the multisig ledger anyway.

This is mostly to keep things consistent and more explicit in the
ledger, and also to simplify some tests.