private tally: add committee date checks
Previously this was done only on the `EncryptedVoteTally` certificate.
So currently the committee end/finish date are not used for private
voteplans.

This reintroduces those checks, but moves them to the moment of
decrypting the tally. This is not exactly the same behaviour as before,
since previously we were only verifying that the decrypted shares were
submitted after the committee start date (because otherwise the tally
was None, so there is nothing to decrypt).

Also, add tests for these cases, since we were only testing the public
case.