Incremental tally (#745)
* vote manager implement incremental tally

* vote/ledger.rs: incremental tally needs token distribution

* VoteProposalStatus: remove ballots and Option

since the Tally is always present, we don't need Option anymore

* VoteTally: remove total stake from Private

since it simplifies the ProposalManager constructor and it can be
obtained easily at the end

* ledger: vote cast now needs token distribution

because the tally is incremental, we need to know the stake

* rename finalize_private_tally to private_tally

* vote manager fix imports

* derive Clone for TokenDistribution

* remove incorret tally test

since the tally is done incrementally, each vote tests that

* remove encrypted tally

* incremental: handle invalid vote error

* vote manager tests fix imports

* move token_distribution out of the VotePlanManager

I think in the future we may want to move it inside again, but for the
time being this simplifies the ledger initialization, since otherwise
the order of MintToken and VotePlan fragments would matter, which is OK
but IMHO a bit fragile.

* reduce imhamt cloning in token distribution

by using normal borrows, at the expense of making the tests slightly
more verbose

* move function argument to the end

* remove extra confirm transaction in test

a remnant from the moment I deleted the EncryptedVoteTally cert

* remove encrypted vote tally from benchmark

* fill fragment tag gaps

* private tally: add committee date checks

Previously this was done only on the `EncryptedVoteTally` certificate.
So currently the committee end/finish date are not used for private
voteplans.

This reintroduces those checks, but moves them to the moment of
decrypting the tally. This is not exactly the same behaviour as before,
since previously we were only verifying that the decrypted shares were
submitted after the committee start date (because otherwise the tally
was None, so there is nothing to decrypt).

Also, add tests for these cases, since we were only testing the public
case.

* remove rayon from dependencies

* replace TallyAlreadyDecrypted for unreachable!

since the vote/tally date validations should account for it

* reject votes with zero tokens

* add missing initial tokens in tests

* add rayon to dev-dependencies for the bench

* add vote_manager_zero_tokens_vote_should_fail

* remove zero voting power vote cast from decrypt_tally panic test

it's not possible anymore to cast a vote with 0 voting power, and there
is already a test for that, so asserting that it is an error here
wouldn't add that much information, so for now, just test with an empty
tally.